+++ apps/test-suite/TestModules.js
@@ -70,6 +70,7 @@ export function getTestModules() {
     require('./tests/FirebaseJSSDK'),
     require('./tests/ImageManipulator'),
     require('./tests/Clipboard'),
+    require('./tests/Fetch'),
     optionalRequire(() => require('./tests/SQLiteLegacy'))
   );